include iterator code
Until now the iterator was directly user triggerd and
only one node distance can be passed. The direct consequence
was that a wrapper script (pyhton) must be used to generate
the required range. This was fine, except for the random number
generator. This was seeded at every start via /dev/urandom. In times
where no IO was triggered the entropy pool was really empty.

The newly introduced "inline" iterator disarm this behavior.